# StackApp/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PerkOS Stack API",
    "version": "2.0.0",
    "description": "x402 facilitator (verify and settle stablecoin payments on 16 EVM networks), ERC-8004 agent identity registration, and agent accounts. Full narrative reference at /llms.txt, authentication at /auth.md.",
    "contact": { "name": "PerkOS", "url": "https://perkos.xyz", "email": "contact@perkos.xyz" }
  },
  "servers": [{ "url": "https://stack.perkos.xyz" }],
  "components": {
    "securitySchemes": {
      "apiKey": { "type": "apiKey", "in": "header", "name": "X-API-Key", "description": "sk_perkos_… from POST /api/v2/agents/register. Also accepted as Authorization: Bearer." },
      "walletSignature": { "type": "apiKey", "in": "header", "name": "X-Wallet-Signature", "description": "EIP-191 signature over 'PerkOS Sponsor Wallet Access {timestamp}', with X-Wallet-Timestamp and X-Wallet-Address." }
    },
    "schemas": {
...
      "X402Request": {
        "type": "object",
        "required": ["x402Version", "paymentPayload", "paymentRequirements"],
        "properties": {
          "x402Version": { "type": "integer", "enum": [1, 2] },
          "paymentPayload": { "type": "object", "properties": { "scheme": { "type": "string" }, "network": { "type": "string" }, "payload": { "type": "object", "properties": { "signature": { "type": "string" }, "authorization": { "type": "object", "properties": { "from": { "type": "string" }, "to": { "type": "string" }, "value": { "type": "string" }, "validAfter": { "type": "string" }, "validBefore": { "type": "string" }, "nonce": { "type": "string" } } } } } } },
          "paymentRequirements": { "type": "object", "properties": { "scheme": { "type": "string" }, "network": { "type": "string" }, "amount": { "type": "string" }, "asset": { "type": "string" }, "payTo": { "type": "string" }, "resource": {}, "maxTimeoutSeconds": { "type": "integer" }, "extra": { "type": "object" } } }
        }
      },
      "VerifyResponse": { "type": "object", "properties": { "isValid": { "type": "boolean" }, "invalidReason": { "type": ["string", "null"] }, "payer": { "type": ["string", "null"] } } },
...
      "Error": { "type": "object", "properties": { "error": { "type": "string" } } }
    }
  },
  "paths": {
...
    "/api/v2/x402/verify": {
      "post": { "operationId": "x402Verify", "summary": "Verify an x402 payment payload against requirements", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/X402Request" } } } }, "responses": { "200": { "description": "Verification result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VerifyResponse" } } } }, "400": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } } } }
    },
...
    "/api/v2/agents/onboard": {
      "post": { "operationId": "prepareErc8004Registration", "summary": "Unsigned ERC-8004 registration calldata for a network", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["network"], "properties": { "network": { "type": "string", "example": "base" }, "tokenURI": { "type": "string" }, "metadata": { "type": "array", "items": { "type": "object" } }, "paymentReceiver": { "type": "string" }, "agentId": { "type": "string" } } } } } }, "responses": { "200": { "description": "Registration plan with calldata" }, "400": { "description": "Network not agent-ready" } } }
    },
    "/api/v2/agents/discovery": {
      "get": { "operationId": "checkAgentDiscovery", "summary": "Whether an ERC-8004 agent is indexed by 8004scan", "parameters": [{ "name": "chainId", "in": "query", "required": true, "schema": { "type": "integer" } }, { "name": "agentId", "in": "query", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Discovery status" }, "400": { "description": "Missing parameters" } } }
    },
...
  }
}